backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
xmltodict==0.14.1
python-docx==1.1.2
